divs for titles & "new announcement/quiz/topic" buttons (#139)
Titles, main content, and "add new X" buttons are in their own separate divs. This was so that I could keep the titles (in the headers) & buttons (in the footers) of the three boxes static, while the middle section, which has the main content (announcements, quizzes, topics), can still scroll around.